ramanujan/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0
pyyaml>=6.0

# Optional ML libraries (install as needed)
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.0.0
tensorflow>=2.8.0
arch>=5.3.0
hmmlearn>=0.2.7

# For visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0

# For hyperparameter optimization (optional)
optuna>=3.0.0
hyperopt>=0.2.7